Move things around a bit
