+++ Cargo.toml
@@ -44,6 +44,8 @@ stats_alloc = "0.1.8"
 rkyv = { version = "0.7.19", features = ["validation", "smallvec"] }
 hex = "0.4.3"
 hexdump = "0.1.1"
+futures = "0.3.17"
+tokio = { version = "1.13.0", features = ["full"] }
 
 [[bench]]
 name = "set_comparison"
@@ -53,3 +55,4 @@ harness = false
 [[bench]]
 name = "set_ops"
 harness = false
+
